==> Rmd/1_clean_EUCLID_dump.Rmd <==
---
title: "Harvest traits from EUCLID htm for Corymbia species"
author: "AC"
date: "06/08/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

librarian::shelf(tidyverse, readtext, conflicted, rvest, DesiQuintans/desiderata)

conflict_prefer("pluck", "purrr")
conflict_prefer("filter", "dplyr")
```

## Aim:
Harvest Eucalyptus species traits from botanical descriptions ("fact sheets") that are available on the EUCLID website <https://apps.lucidcentral.org/euclid/>. In this document, I will clean up the raw data and get it into a state where I can begin extracting trait data from it.

Extract life-history trait information for the PGLS analysis on Victoria's CCA dataset.

The EUCLID data and corymbia_species.txt were obtained by extracting the urls from the .htm websites for Corymbia species <https://apps.lucidcentral.org/euclid/text/entities/index_corymbia.htm>.


# 01. Dumping the website

There is a webpage of all Corymbia species on EUCLID with links to their fact sheets at <https://apps.lucidcentral.org/euclid/text/entities/index.htm>. I downloaded the code for these webpages and used a regular expression to grab just the URLs to the fact sheets. Those URLs are in a plain text file in `raw_data/corymbia_species.txt`:

  `https://apps.lucidcentral.org/euclid/text/entities/corymbia_abbreviata.htm`
  `https://apps.lucidcentral.org/euclid/text/entities/corymbia_abergiana.htm`
  `https://apps.lucidcentral.org/euclid/text/entities/corymbia_aparrerinja.htm`
  `https://apps.lucidcentral.org/euclid/text/entities/corymbia_arafurica.htm`
  `https://apps.lucidcentral.org/euclid/text/entities/corymbia_arenaria.htm`

Linux command used: `cat EUCLIDSpeciesIndex.htm | tr '"' '\n' | tr "'" '\n' | grep -e '^https://' -e '^http://' -e'^//' | sort | uniq > corymbia_species.txt`

Then I used `wget` <https://www.gnu.org/software/wget/> to download every URL in the list:

    wget -i "corymbia_species.txt"
    
The downloaded HTML files were moved to `raw_data/EUCLID_dump_corymbia` and `raw_data/EUCLID_dump_angophora`:

  `raw_data/EUCLID_dump_angophora/angophora_bakeri_subsp._bakeri.htm`
  `raw_data/EUCLID_dump_angophora/angophora_bakeri_subsp._crassifolia.htm`
  `raw_data/EUCLID_dump_angophora/angophora_costata_subsp._costata.htm`
  `raw_data/EUCLID_dump_angophora/angophora_costata_subsp._euryphylla.htm`

The files can then be held in a big dataframe for working.

``` {r construct EUCLID database}
raw_pages_corymbia <- 
    # readtext may produce warnings if the files being imported are *.htm. 
    # That's okay because it imports them as plain text, which is what we want.
    readtext(list.files("../raw_data/EUCLID_dump_corymbia/", full.names = TRUE)) %>% 
    rename(name = doc_id, html = text) %>% 
    # The name field can be used to grab the species/subspecies later.
    mutate(name = str_remove(name, "\\.htm"),
           name = str_replace_all(name, "_", " ")) %>% 
    # Remove soft linebreaks and multiple spaces from HTML to make it easier to 
    # see what I am retrieving. 
    # I also replace the em dashes, which failed to encode properly, with --.
    # I also replace <br /> with a text equivalent ( -BREAK- ) because it is 
    # used in the HTML to separate parts of the description.
    mutate(html = str_replace_all(html, "\\\n", " "),
           html = str_replace_all(html, "\\s+", " "),
           html = str_replace_all(html, "â€“", "--"),
           html = str_replace_all(html, "Â°", "°"),
           html = str_replace_all(html, "<br/>|<br />|<BR>", " -BREAK- ")) %>% 
    # The HTML (as plain text) needs to be converted for use with rvest. It must
    # be held in a list column because rvest outputs a 2-element list.
    group_by(name) %>%
    mutate(html = read_html(html) %>% list()) %>%
    ungroup() %>%
    identity()

glimpse(raw_pages_corymbia)

```

# 02. Isolating the sections of the fact sheet

It's lovely that the different sections of the fact sheet are cleanly separated from each other in the HTML.

```{r}
# Remove the linebreak placeholders and clean up extra spaces.
clean_text <- function(text) {
    text %>% 
        str_remove_all("-BREAK-") %>% 
        str_trim()
}


get_classification <- function(html) {
    html %>%
        pluck(1) %>%   # The HTML is in a list column and must be plucked out.
        html_node("div#classification") %>%
        html_node("div.content") %>%
        html_text() %>%
        clean_text()
}


get_nomenclature <- function(html) {
    html %>% 
        pluck(1) %>%   # The HTML is in a list column and must be plucked out.
        html_node("div#nomenclature") %>% 
        html_node("div.content") %>% 
        html_text() %>% 
        clean_text()
}


get_description <- function(html) {
    html %>% 
        pluck(1) %>%   # The HTML is in a list column and must be plucked out.
        html_node("div#description") %>% 
        html_node("div.content") %>% 
        html_text()
        # No clean_text() here because I need the linebreak placeholders later.
}
```

The output is like:

```{r}
get_classification(raw_pages_corymbia$html[1])
get_nomenclature(raw_pages_corymbia$html[1])
get_description(raw_pages_corymbia$html[1])
```

This can be applied to each entry now:

```{r}
raw_sections_corymbia <- 
    raw_pages_corymbia %>% 
    group_by(name) %>% 
    summarise_at("html", 
                 list(~ get_classification(.),
                      ~ get_nomenclature(.),
                      ~ get_description(.)
                      )
                 ) %>% 
    ungroup()
```

# 03. Splitting the description

Each part of the description that they have asked for (growth form, adult leaves, fruit, seed, and inflorescence) can be further split off.

```{r}
get_growthform <- function(desc) {
    desc %>% 
        str_extract("^.*?(-BREAK-|$)") %>% 
        clean_text()
}


get_adultleaves <- function(desc) {
    out <- 
        desc %>% 
        # Sometimes "Adult leaves" is inside the sentence.
        # "Crown composed entirely of Adult leaves which are..."
        str_extract("(Adult|Crown|crown|Mature|mature) leaves.*?(-BREAK-|$)") %>% 
        clean_text()
    
    # I try to search for "Adult leaves" first because for some species, the 
    # lowercase "adult leaves" comes earlier in the description and is 
    # wrongly retrieved instead.
    if(is.na(out)) {
        out <- 
            desc %>% 
            # Sometimes "Adult leaves" is inside the sentence.
            # "Crown composed entirely of Adult leaves which are..."
            str_extract("adult leaves.*?(-BREAK-|$)") %>% 
            clean_text()
    }
    
    # If the retrieved section does not have a number in it, it is probably 
    # because the juvenile leaves are the ones being referred to when the adult
    # leaves are described.
    if(str_detect(out, "cm (long|wide)") == FALSE | is.na(out)) {
        out <- NA_character_
    }
    
    return(out)
}


get_fruit <- function(desc) {
    desc %>% 
        str_extract("-BREAK- Fruit.*?(-BREAK-|$)") %>% 
        clean_text()
}


get_seeds <- function(desc) {
    desc %>% 
        str_extract("Seed.*?(-BREAK-|$)") %>% 
        clean_text()
}


get_inflorescence <- function(desc) {
    desc %>% 
        str_extract("(-BREAK-)? Inflorescence.*?(-BREAK-|$)") %>% 
        clean_text()
}
```

The output looks like this:

```{r}
get_growthform(raw_sections_corymbia$get_description[1])
get_adultleaves(raw_sections_corymbia$get_description[799])
get_fruit(raw_sections_corymbia$get_description[1])
get_seeds(raw_sections_corymbia$get_description[1])
get_inflorescence(raw_sections_corymbia$get_description[1])
```

These columns can be retrieved too. This dataframe is now ready for the trait extraction process.

```{r}
sections_clean <- 
    raw_sections_corymbia %>% 
    group_by(name) %>% 
    mutate_at("get_description", 
                 list(~ get_growthform(.),
                      ~ get_adultleaves(.),
                      ~ get_fruit(.),
                      ~ get_seeds(.),
                      ~ get_inflorescence(.))) %>% 
    ungroup() %>% 
    select(name, get_classification, get_growthform:get_inflorescence, 
           everything()) %>% 
    glimpse()
```

# 04. Manually fix adult leaf section

There are 15 species that fail to split properly because they do not have a proper section for adult leaves. This is because the adult leaves are similar enough to the juvenile ones that the description simply points to what was already written for the juveniles. I will manually insert those juvenile leaf descriptions into the space for adult leaves.

```{r}
rows_with_na(sections_clean) %>% pluck("name")
```

```{r}
abbreviata_leaves <-
  uw("Crown of stiff, juvenile leaves, opposite, sessile or on very short petioles (0–0.5 cm); blade cordate to broadly lanceolate, 7–23.5 cm long, 3–11 cm wide, undulate, base usually amplexicaul or lobed, rarely rounded, margin entire, apex pointed, concolorous or slightly discolorous, dull, green but yellowish near main veins, glabrous (rarely a few sparse bristle-glands at the base of the lamina), side-veins at a wide angle to the midrib and widely spaced, and looping to form intramarginal vein, reticulation very dense, oil glands apparently absent.")

cadophora_cadophora_leaves <-
  uw("Crown of juvenile leaves which are always opposite and sessile with the leaf-bases joined, thus encircling the stem; blade ovate to lanceolate, each leaf of the joined pair measuring 9.5–25.5 cm long, 3–8.6 cm wide, margin entire or slightly irregular, apex pointed or rounded, concolorous, dull, green to grey-green, smooth, side-veins at a wide angle to the midrib (penniveined), reticulation very dense, intramarginal vein present and close to the margin or confluent with the margin, oil glands apparently absent; setae sometimes present on midrib close to stem, and also sometimes present on new growth before leaf expansion.")

cadophora_polychroma_leaves <- 
  uw("Crown of juvenile leaves which are always opposite and sessile with the leaf-bases joined encircling the stem; blade ovate to oblong, each leaf of the joined pair measuring 9–20.5 cm long, 4.9–10.2 cm wide, undulate, margin entire or slightly irregular, apex pointed or rounded, concolorous or weakly discolorous, dull, green to grey-green, smooth, side-veins at a wide angle to the midrib (penniveined), reticulation very dense, intramarginal vein present and close to the margin or confluent with the margin, oil glands apparently absent; setae sometimes present on midrib close to stem present and on new growth before leaf expansion.")

chartacea_leaves <- 
  uw("Crown of juvenile leaves, opposite, sessile or on very short petioles (0–0.2 cm); blade cordate to broadly elliptical, (4.5)6–16 cm long, 3–8.6 cm wide, base amplexicaul, margin entire, apex pointed or rounded, concolorous, dull, yellow-green to light green to grey-green, smooth, side-veins at a wide angle to the midrib (penniveined), reticulation very dense, intramarginal vein present, close to the margin, oil glands apparently absent, leaf surface weakly scabrid with remains of bristle-glands (setae) or leaf smooth.")

ferruginea_ferruginea_leaves <- 
  uw("Crown of juvenile leaves, opposite, rarely a few sub-opposite, sessile or on very short petioles (0–0.8 cm); blade broadly lanceolate to ovate or elliptical, 6.3–18 cm long, 3–9 cm wide, undulate, base usually amplexicaul or lobed, margin entire, apex pointed or rounded and apiculate, concolorous, dull, green but brown near main veins, scabrid with short rusty bristle-glands, side-veins at a wide angle to the midrib and widely spaced, looping to form intramarginal vein, reticulation very dense, oil glands apparently absent.")

pachycarpa_leaves <- 
  uw("Crown of juvenile leaves, opposite, sessile; blade cordate to ovate to lanceolate, 4–12 cm long, 1.7–5 cm wide, base amplexicaul, margin entire, apex pointed or rounded and apiculate, concolorous, dull, yellow-green to light green to grey-green, smooth or slightly scabrid with the remains of short bristle-glands, side-veins at a wide angle to the midrib, reticulation very dense, intramarginal vein present, oil glands scattered island or apparently absent.")

scabrida_leaves <- 
  uw("Crown entirely of scabrid juvenile and intermediate leaves, predominantly alternately arranged, rarely a few pairs sub-opposite, petiole 0.8–1.7 cm long; blade usually lanceolate, sometimes ovate-elliptic to ovate, (5)6.5–12.5 cm long, (1.3)2–3.5(4.8) cm wide, flat or undulate, base of most leaves peltate, rarely rounded to tapering to petiole, apex rounded or pointed, margin entire, concolorous, dull, grey-green, penniveined, densely to very densely reticulate, intramarginal vein parallel to and just within margin, oil glands island or obscure.")

sphaerica_leaves <- 
  uw("Crown of juvenile leaves, opposite, sessile or very shortly petiolate (petioles to only 0.1 cm long); blade cordate to almost lanceolate, 3–7(10) cm long, 1.5–4.5 cm wide, base amplexicaul, margin entire, apex pointed or rounded and apiculate, concolorous, dull, light green to grey-green or blue-green, side-veins at a wide angle to the midrib (penniveined), reticulation very dense, intramarginal vein present, oil glands apparently absent.")

aspera_leaves <- 
  uw("Crown consists of juvenile leaves, opposite, sessile or with petioles to 0.3 cm long; blade cordate to ovate or rarely ± elliptical, 1.5–6.8 cm long, 0.7–3.5 cm wide, base amplexicaul to rounded, margin entire, apex rounded, concolorous, dull, green to grey-green, usually setose to scabrid with bristle-glands, side-veins at more than 45° to midrib, reticulation dense to very dense, intramarginal vein present, oil glands not visible.")

cadophora_pliantha_leaves <- 
  uw("Crown of juvenile leaves which are always opposite and sessile with the leaf-bases joined, thus encircling the stem; blade ovate to lanceolate, each leaf of the joined pair measuring 7–17.5 cm long, 2.8–7.8 cm wide, margin entire or slightly irregular, apex pointed or rounded, concolorous, dull, green to grey-green, smooth, side-veins at a wide angle to the midrib (penniveined), reticulation very dense, intramarginal vein present and close to the margin or confluent with the margin, oil glands apparently absent; setae sometimes present on midrib close to stem present and on new growth before leaf expansion.")

candida_leaves <- 
  uw("Crown of juvenile opposite, sessile or with petioles to 0.4 cm long; blade cordate, rarely lanceolate, 3–8 cm long, 0.8–2.7 cm wide, base amplexicaul, margin entire, apex pointed, concolorous, dull, grey-green to pale green, usually scabrid but older leaves may be weathered to ± smooth, side-veins at more than 45° to midrib, reticulation very dense, intramarginal vein present, oil glands absent.")

dunlopiana_leaves <- 
  uw("Crown of juvenile leaves, opposite, sessile; blade cordate to oblong-lanceolate, 3.5–12 cm long, 2–5.7 cm wide, base amplexicaul, margin entire, apex usually pointed, flat or undulate, concolorous, dull, yellow-green to light green to grey-green, prominently setose with bristle-glands, especially on margins and main veins (rarely the leaves ± glabrous in some populations), side-veins at a wide angle to the midrib (penniveined), reticulation dense to very dense, intramarginal vein present close to margin, oil glands apparently absent from the venation pattern.")

ferruginea_stypophylla_leaves <- 
  uw("Crown of juvenile leaves, opposite, rarely a few sub-opposite, petiolate, petioles (0.5)0.8–2.6 cm long; blade usually broadly lanceolate to ovate, (5.4)7–21.5 cm long, 2–6(7.5) cm wide, undulate, base tapering to petiole or truncate to rounded, sometimes lobed, margin usually entire, apex pointed or rounded and ± apiculate, concolorous, dull or weathering slightly glossy, green but brown near main veins, scabrid with short rusty bristle-glands (in some populations scabridity obvious only on newest expandings leaves), side-veins at a wide angle to the midrib and widely spaced, looping to form intramarginal vein, reticulation very dense, oil glands apparently absent. reticulation very dense, intramarginal vein present, oil glands apparently absent from the venation pattern.")

papillosa_leaves <- 
  uw("Crown of thin-textured juvenile leaves, opposite, sessile or with petioles to 0.8 cm long; blade oblong to elliptical, (3.5)4–12.2 cm long, 2–5.5 cm wide, base amplexicaul or lobed, margin entire, apex pointed or rounded and apiculate, concolorous, dull, yellow-green to light green to grey-green, scabrid with dense short setae more or less even in length, side-veins at a wide angle to the midrib (penniveined), reticulation very dense, intramarginal vein present, close to the margin, oil glands apparently absent from the venation pattern.")

setosa_leaves <- 
  uw("Crown of juvenile leaves, opposite, sessile or very shortly petiolate (to only 0.2 cm long); blade cordate or rarely almost elliptical-oblong, (1.5)2.7–7.7 cm long, 1–5 cm wide, base amplexicaul, margin entire, apex pointed or rounded and apiculate, concolorous, dull, light green to grey-green or yellow-green, smooth or slightly scabrid with the remains of short bristle-glands, side-veins at a wide angle to the midrib (penniveined), reticulation very dense, intramarginal vein present, oil glands apparently absent.")

fixed_sections <- 
    sections_clean %>% 
    mutate(get_adultleaves = 
               case_when(name == "corymbia abbreviata" ~ abbreviata_leaves,
                         name == "corymbia cadophora subsp. cadophora" ~ cadophora_cadophora_leaves,
                         name == "corymbia cadophora subsp. polychroma" ~ cadophora_polychroma_leaves,
                         name == "corymbia chartacea" ~ chartacea_leaves,
                         name == "corymbia ferruginea subsp. ferruginea" ~ ferruginea_ferruginea_leaves,
                         name == "corymbia pachycarpa" ~ pachycarpa_leaves,
                         name == "corymbia scabrida" ~ scabrida_leaves,
                         name == "corymbia sphaerica" ~ sphaerica_leaves,
                         name == "corymbia aspera" ~ aspera_leaves,
                         name == "corymbia cadophora subsp. pliantha" ~ cadophora_pliantha_leaves,
                         name == "corymbia candida" ~ candida_leaves,
                         name == "corymbia dunlopiana" ~ dunlopiana_leaves,
                         name == "corymbia ferruginea subsp. stypophylla" ~ ferruginea_stypophylla_leaves,
                         name == "corymbia papillosa" ~ papillosa_leaves,
                         name == "corymbia setosa" ~ setosa_leaves,
                         TRUE ~ get_adultleaves))

rows_with_na(fixed_sections)
```

# Output

```{r}
write_rds(fixed_sections, "../intermediate/factsheet_as_sections.rds")
write_csv(fixed_sections, "../intermediate/factsheet_as_sections.csv")
```
